requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0